팝업 수정

+++ MESDTO/StdStepInfoVO.cs
@@ -11,7 +11,7 @@ namespace MESDTO
         public string STD_STEP_ID { get; set; }
         public string STD_STEP_NAME { get; set; }
         public int STEP_TAT { get; set; }
-        public int STEP_YIELD { get; set; }
+        public decimal STEP_YIELD { get; set; }
         public int STEP_SETUP { get; set; }
         public string MODIFIER { get; set; }
         public DateTime MODIFIER_DATE { get; set; }
